+++ src/index.tsx
@@ -37,19 +37,10 @@ const router = createBrowserRouter([
 ]);
 
 root.render(
+  // WARNING: StrictMode causes double renders, which causes problems
+  // during development when trying to open previously used serial ports
+  // and also with loading other things from local storage
   // <React.StrictMode>
     <RouterProvider router={router} />
   // </React.StrictMode>
 );
-
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://cra.link/PWA
-// serviceWorkerRegistration.unregister();
-// serviceWorkerRegistration.register();
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();