+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Last serial port details and connection state are remember across App reloads, and NinjaTerm can automatically reconnect.
+
 ### Changed
 
 - Google Analytics is now only initialized in production builds, to prevent things like Playwright e2e tests from spamming GA and skewing data.